+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Typo in cache constant.
+
+### Added
+
+- Settings page.
 - Added github actions
 
 ## [1.1.0] 2024-02-01